feat: add composable + component approach
