+++ playground/components.d.ts
@@ -9,6 +9,7 @@ declare module 'vue' {
   export interface GlobalComponents {
     AkuAku: typeof import('./src/components/AkuAku.vue')['default']
     DirectivesDemo: typeof import('./src/components/DirectivesDemo.vue')['default']
+    FboCube: typeof import('./src/components/fboCube.vue')['default']
     Gltf: typeof import('./src/components/gltf/index.vue')['default']
     ModelsDemo: typeof import('./src/components/ModelsDemo.vue')['default']
     RouterLink: typeof import('vue-router')['RouterLink']
