+++ src/core/abstractions/useFbo/component.vue
@@ -1,0 +1,16 @@
+<script setup lang="ts">
+import type { FboOptions } from '.'
+import { useFBO } from '.'
+
+const props = withDefaults(defineProps<FboOptions>(), {
+  depth: false,
+  samples: 0,
+  settings: undefined,
+})
+
+const target = useFBO(props)
+
+defineExpose({
+  value: target,
+})
+</script>
